requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
transformers
accelerate
sentence-transformers
pandas
numpy
scikit-learn
matplotlib
seaborn
Pillow
requests
tqdm
jupyter
ipywidgets
